KB and RAG/RAG/requirements.txt
pip
setuptools
wheel

# LangChain dependencies
langchain
langchain-core
langchain-groq
langchain-experimental
langchain-community
langchain-pinecone

# Pinecone
pinecone-client

# Web Framework
fastapi
uvicorn